fix(chat-custom-width): Correct unit and initial prompt sync state issues
- Change percentage values to REM units where appropriate
- Fix the initial state of the `textarea-sync` switch to ensure it starts as expected

Changes summary:
- Corrected unit inconsistencies by changing percentage values to REM units where appropriate, ensuring consistent sizing across different screen sizes and font sizes.
- Addressed an issue with the initial state of the `textarea-sync` switch, ensuring it starts in the expected default state.
