+++ src/js/content.js
@@ -1,6 +1,5 @@
 import './app/floatingBtn'
 import './app/mainFonts'
-// import { init as initAssets } from './app/mainAssets'
-import { init as initAssets } from './app/mainAssetsUpgrade'
+import { init as initAssets } from './app/mainAssets'
 
 initAssets()
